Evolution of Object Factory beginning with online tutorial
// objectCreate_onlineCodeBefore_objectCreate.js [file content includes the Object.create() in the evolution to an object factory - not the original intent and I did not change the file name.]
Begins with material from:
// http://www.htmlgoodies.com/beyond/javascript/object.create-the-new-way-to-create-objects-in-javascript.html
and evolves to using SPA book Chapter 2 page 39 object factory.
Result is on;
// https://jsbin.com/codida/edit?js,console,output
the factory allows single line instance creation of objects for normally unique values for the instance.
For cars, the brand and color were used. 
An explanation of a property descriptor is included; both for a data descriptor and an accessor descriptor -- extracted from the online reference and clarified with explicit examples.