Javascript Object.create() to Emulate Constructor
// Object.createAsConstuctor.js
In the article JavaScript: How to Use Object.create() to Emulate Constructor? at http://xahlee.info/js/javascript_object_create_emulating_constructor.html,
The instance is created with a long descriptor --  an in real life they are even longer. I would prefer a one line instantiation, e.g.,
     var o2 = fcn(4);
The attached file rewrites the cited article to do just that. The original descriptor is returned from a function that is invoked as the second argument to Object.create(), instead of having the entire multi-line object repeated as the second argument for each instance.
The approach is easily extended to multiple arguments in the descriptor.
This approach is applicable to get rid of 'new' and 'this' in such exercises as in codeAcademy Javascript I - Lesson 21 Custom Constuctors.